+++ ui/src/components/actions-menu/QasActionsMenu.vue
@@ -1,28 +1,5 @@
 <template>
   <div v-if="hasList" class="inline-block qas-actions-menu" data-cy="actions-menu">
-    <!-- <div v-if="hasActions" class="qas-actions-menu" data-cy="actions-menu"> -->
-    <!-- <component :is="component.is" v-bind="component.props" variant="tertiary" @click.stop.prevent>
-      <q-list v-if="isBtnDropdown" data-cy="actions-menu-list">
-        <slot v-for="(item, key) in actions" :item="item" :name="key">
-          <q-item v-bind="item.props" :key="key" clickable data-cy="actions-menu-list-item" @click="onClick(item)">
-            <q-item-section avatar>
-              <q-icon :name="item.icon" />
-            </q-item-section>
-
-            <q-item-section>
-              <div>
-                {{ item.label }}
-              </div>
-            </q-item-section>
-          </q-item>
-        </slot>
-      </q-list>
-
-      <q-tooltip v-if="hasTooltip" class="text-caption">
-        {{ tooltipLabel }}
-      </q-tooltip>
-    </component> -->
-
     <qas-btn-dropdown v-bind="btnDropdownProps">
       <q-list data-cy="actions-menu-list">
         <slot v-for="(item, key) in formattedList.dropdownList" :item="item" :name="key">
@@ -86,11 +63,6 @@ const props = defineProps({
     type: Object
   },
 
-  dropdownIcon: {
-    default: 'sym_r_more_vert',
-    type: String
-  },
-
   list: {
     default: () => ({}),
     type: Object
@@ -120,12 +92,30 @@ const screen = useScreen()
 
 const { deleteBtnProps, hasDelete } = useDelete()
 
-const listSize = computed(() => Object.keys(fullList.value).length)
 const hasSplitName = computed(() => !!props.splitName)
-const hasSplit = computed(() => hasSplitName.value && listSize.value > SPLIT_SIZE)
-const isSingle = computed(() => Object.keys(fullList.value).length === 1)
 const hasList = computed(() => !!Object.keys(fullList.value).length)
 
+/**
+ * Tamanho total da lista, considerando ação de deletar caso exista.
+ */
+const listSize = computed(() => Object.keys(fullList.value).length)
+
+/**
+ * Só existe split caso tenha a propriedade "splitName" e o tamanho da lista seja
+ * maior que o tamanho permitido no SPLIT_SIZE (no caso 2).
+ */
+const hasSplit = computed(() => hasSplitName.value && listSize.value > SPLIT_SIZE)
+
+/**
+ * Quando existe apenas i item na lista, neste caso é mostrado o botão direto,
+ * mesmo que não tenha a propriedade "splitName".
+ */
+const isSingle = computed(() => Object.keys(fullList.value).length === 1)
+
+/**
+ * Lista englobando as que vem por propriedade "list" mergeadas com a do
+ * delete, caso exista.
+ */
 const fullList = computed(() => {
   return {
     ...props.list,
@@ -133,9 +123,20 @@ const fullList = computed(() => {
   }
 })
 
-const primaryKey1 = computed(() => {
+/**
+ * chave do primaria do objeto, é considerado sendo a splitName caso exista no
+ * objeto, porém pode acontecer deste item deixar de existir por algum motivo
+ * então consideramos a chave primaria sendo o primeiro item da computada fullList.
+ */
+const primaryKey = computed(() => {
   return props.splitName in fullList.value ? props.splitName : Object.keys(fullList.value)?.[0]
 })
+
+/**
+ * Computadas para controle do tooltip
+ */
+const hasTooltip = computed(() => !props.useLabel && props.useTooltip && isSingle.value)
+const tooltipLabel = computed(() => fullList.value[primaryKey.value]?.label)
 
 const btnDropdownProps = computed(() => {
   return {
@@ -146,12 +147,19 @@ const btnDropdownProps = computed(() => {
 })
 
 const primaryButtonProps = computed(() => {
-  const buttonProps = fullList.value[primaryKey1.value] || {}
+  const buttonProps = fullList.value[primaryKey.value] || {}
 
   return {
+    /**
+     * Caso seja "isSingle" e tenha a opção de deletar, significa que único botão
+     * que existe é o de deletar, então a cor precisa ser "DEFAULT_COLOR", em todos
+     * os outros cenários, é primary.
+     */
     color: isSingle.value && hasDelete.value ? DEFAULT_COLOR : 'primary',
 
     ...buttonProps,
+
+    label: getLabel(buttonProps),
     onClick: () => setClickHandler(buttonProps)
   }
 })
@@ -166,17 +174,19 @@ const formattedList = computed(() => {
   const payload = { dropdownList: {}, buttonsList: {} }
 
   /**
-   * Cenário onde não é utilizado com split e não tem apenas 1 item, sendo o botão de "opções",
-   * onde será usado em casos como dentro de cards, etc.
+   * Cenário onde não é utilizado com split e tenha mais de 1 item na lista,
+   * neste caso, ele se torna o botão de "opções" onde será usado em casos como
+   * dentro de cards, etc.
    */
   if ((!hasSplitName.value || screen.isSmall) && !isSingle.value) {
     payload.buttonsList.options = {
       color: DEFAULT_COLOR,
       iconRight: 'sym_r_more_vert',
-      label: getLabel({ label: 'Opções' }),
       useLabelOnSmallScreen: false,
 
-      ...props.buttonProps
+      ...props.buttonProps,
+
+      label: getLabel({ label: 'Opções' })
     }
 
     payload.dropdownList = buttonsPropsList
@@ -184,17 +194,13 @@ const formattedList = computed(() => {
     return payload
   }
 
-  const primaryKey = primaryKey1.value
-  const primaryButton = { ...primaryButtonProps.value }
-  // const primaryButton = buttonsPropsList[primaryKey]
-
   /**
    * Cenário para quando existir apenas 1 item na lista.
    */
   if (isSingle.value) {
     payload.buttonsList = {
-      [primaryKey]: {
-        ...primaryButton
+      [primaryKey.value]: {
+        ...primaryButtonProps.value
       }
     }
 
@@ -205,14 +211,13 @@ const formattedList = computed(() => {
    * Cenário onde ficará um botão ao lado do outro "Ação primaria | Ação secundaria".
    */
   if (listSize.value === SPLIT_SIZE) {
-    const secondaryKey = Object.keys(buttonsPropsList).find(key => key !== primaryKey)
-    console.log('TCL: formattedList -> secondaryKey', secondaryKey)
+    const secondaryKey = Object.keys(buttonsPropsList).find(key => key !== primaryKey.value)
     const secondaryButton = buttonsPropsList[secondaryKey]
 
     payload.buttonsList = {
-      [primaryKey]: {
-        ...primaryButton,
-        label: getLabel(primaryButton)
+      [primaryKey.value]: {
+        ...primaryButtonProps.value,
+        label: getLabel(primaryButtonProps.value)
       },
 
       [secondaryKey]: {
@@ -231,23 +236,18 @@ const formattedList = computed(() => {
    */
   if (listSize.value > SPLIT_SIZE) {
     payload.buttonsList = {
-      [primaryKey]: {
-        ...primaryButton,
-        label: getLabel(primaryButton)
+      [primaryKey.value]: {
+        ...primaryButtonProps.value,
+        label: getLabel(primaryButtonProps.value)
       }
     }
 
-    delete buttonsPropsList[primaryKey]
+    delete buttonsPropsList[primaryKey.value]
 
     payload.dropdownList = buttonsPropsList
   }
 
   return payload
-})
-
-const tooltipLabel = computed(() => primaryButtonProps.value?.label)
-const hasTooltip = computed(() => {
-  return !props.useLabel && props.useTooltip && isSingle.value
 })
 
 // functions